Support SF1000+

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE REGION  ( R_REGIONKEY  INTEGER NOT NULL,
                             R_NAME       CHAR(25) NOT NULL,
                             R_COMMENT    VARCHAR(152));
 
-CREATE TABLE PART  ( P_PARTKEY     INTEGER NOT NULL,
+CREATE TABLE PART  ( P_PARTKEY     BIGINT NOT NULL,
                           P_NAME        VARCHAR(55) NOT NULL,
                           P_MFGR        CHAR(25) NOT NULL,
                           P_BRAND       CHAR(10) NOT NULL,
@@ -18,7 +18,7 @@ CREATE TABLE PART  ( P_PARTKEY     INTEGER NOT NULL,
                           P_RETAILPRICE DECIMAL(15,2) NOT NULL,
                           P_COMMENT     VARCHAR(23) NOT NULL);
 
-CREATE TABLE SUPPLIER ( S_SUPPKEY     INTEGER NOT NULL,
+CREATE TABLE SUPPLIER ( S_SUPPKEY     BIGINT NOT NULL,
                              S_NAME        CHAR(25) NOT NULL,
                              S_ADDRESS     VARCHAR(40) NOT NULL,
                              S_NATIONKEY   INTEGER NOT NULL,
@@ -26,13 +26,13 @@ CREATE TABLE SUPPLIER ( S_SUPPKEY     INTEGER NOT NULL,
                              S_ACCTBAL     DECIMAL(15,2) NOT NULL,
                              S_COMMENT     VARCHAR(101) NOT NULL);
 
-CREATE TABLE PARTSUPP ( PS_PARTKEY     INTEGER NOT NULL,
-                             PS_SUPPKEY     INTEGER NOT NULL,
-                             PS_AVAILQTY    INTEGER NOT NULL,
+CREATE TABLE PARTSUPP ( PS_PARTKEY     BIGINT NOT NULL,
+                             PS_SUPPKEY     BIGINT NOT NULL,
+                             PS_AVAILQTY    BIGINT NOT NULL,
                              PS_SUPPLYCOST  DECIMAL(15,2)  NOT NULL,
                              PS_COMMENT     VARCHAR(199) NOT NULL);
 
-CREATE TABLE CUSTOMER ( C_CUSTKEY     INTEGER NOT NULL,
+CREATE TABLE CUSTOMER ( C_CUSTKEY     BIGINT NOT NULL,
                              C_NAME        VARCHAR(25) NOT NULL,
                              C_ADDRESS     VARCHAR(40) NOT NULL,
                              C_NATIONKEY   INTEGER NOT NULL,
@@ -41,8 +41,8 @@ CREATE TABLE CUSTOMER ( C_CUSTKEY     INTEGER NOT NULL,
                              C_MKTSEGMENT  CHAR(10) NOT NULL,
                              C_COMMENT     VARCHAR(117) NOT NULL);
 
-CREATE TABLE ORDERS  ( O_ORDERKEY       INTEGER NOT NULL,
-                           O_CUSTKEY        INTEGER NOT NULL,
+CREATE TABLE ORDERS  ( O_ORDERKEY       BIGINT NOT NULL,
+                           O_CUSTKEY        BIGINT NOT NULL,
                            O_ORDERSTATUS    CHAR(1) NOT NULL,
                            O_TOTALPRICE     DECIMAL(15,2) NOT NULL,
                            O_ORDERDATE      DATE NOT NULL,
@@ -51,10 +51,10 @@ CREATE TABLE ORDERS  ( O_ORDERKEY       INTEGER NOT NULL,
                            O_SHIPPRIORITY   INTEGER NOT NULL,
                            O_COMMENT        VARCHAR(79) NOT NULL);
 
-CREATE TABLE LINEITEM ( L_ORDERKEY    INTEGER NOT NULL,
-                             L_PARTKEY     INTEGER NOT NULL,
-                             L_SUPPKEY     INTEGER NOT NULL,
-                             L_LINENUMBER  INTEGER NOT NULL,
+CREATE TABLE LINEITEM ( L_ORDERKEY    BIGINT NOT NULL,
+                             L_PARTKEY     BIGINT NOT NULL,
+                             L_SUPPKEY     BIGINT NOT NULL,
+                             L_LINENUMBER  BIGINT NOT NULL,
                              L_QUANTITY    DECIMAL(15,2) NOT NULL,
                              L_EXTENDEDPRICE  DECIMAL(15,2) NOT NULL,
                              L_DISCOUNT    DECIMAL(15,2) NOT NULL,
